refactor the way languages are processed
This commit changes the way languages are processed in Topiary.

To properly process an input stream, we need two things:
- a grammar, to produce a syntax tree from the input,
- a query file, to format this syntax tree.

Additionally, we use the intermediate notion of `Language`,
which is represented by an enum type internally.

The CLI for Topiary ensures we get at least one of the following
arguments:
- `--language`, to specify the language of the input stream,
- `--query`, to specify which query file to use,
- `--input-file`, to specify a file from which to read the input.

If `--input-file` is missing, Topiary reads from `stdin`.

We follow the following steps to determine the grammar and the query:

First, we determine the language:
- if the `--language` argument is present, use it,
- else if the `--input-file` is present, guess the language from the
  file extension,
- else, use the language specified in the query file.

Then the language determines the grammar(s). Indeed, the same language
can have multiple grammars: for instance OCaml has separate grammars
for the interface (`.mli`) and the implementation (`.ml`) files.
We test all grammars associated with the language, until one succeeds
or all fail.

The query file is then determined as follows:
- if the `--query` argument is present, use it,
- else automatically deduce the query file from the language.

This commit consolidates all operations regarding languages and their
transformation into grammars or query files in the `language.rs` file.