[OCaml] equivalence classes for syntax node types
Some interface files parse correctly with both the interface grammar
and the implementation grammar, but produce syntax trees in which some
node types have changed. For instance, the following piece of code:
```
include Type_intf.Type
```
produces this tree when parsed with the interface grammar:
```
{Node include_module_type (27, 0) - (27, 22)} - Named: true
  {Node include (27, 0) - (27, 7)} - Named: false
  {Node module_type_path (27, 8) - (27, 22)} - Named: true
    {Node extended_module_path (27, 8) - (27, 17)} - Named: true
      {Node module_name (27, 8) - (27, 17)} - Named: true
    {Node . (27, 17) - (27, 18)} - Named: false
    {Node module_type_name (27, 18) - (27, 22)} - Named: true
```
but produces this tree when parsed with the implementation grammar:
```
{Node include_module (27, 0) - (27, 22)} - Named: true
  {Node include (27, 0) - (27, 7)} - Named: false
  {Node module_path (27, 8) - (27, 22)} - Named: true
    {Node module_path (27, 8) - (27, 17)} - Named: true
      {Node module_name (27, 8) - (27, 17)} - Named: true
    {Node . (27, 17) - (27, 18)} - Named: false
    {Node module_name (27, 18) - (27, 22)} - Named: true
```

However, we want both trees to be formatted the same way:
we don't want the way we've chosen to input the file in Topiary to
change the output result.

To ensure this property, this commit introduces the notion of equivalence
classes over node types in a query file. Simply put, if two node types
belong to the same equivalence class, then they must appear in the same
queries. This guarantees that the syntax trees produced by two different
grammars will be formatted the same way.

So far, the following equivalence classes have been identified for OCaml:
- `include_module`, `include_module_type`
- `module_type_path`, `module_path`, `extended_module_path`
- `module_type_name`, `module_name`

The equivalence classes are not enforced yet, it's up to the maintainers
to respect them. This could change in the future.