[OCaml] `.mli` file support
This commit adds support for interface files in OCaml:
There are now 3 languages: `Ocaml`, `OcamlInterface`, and
`OcamlImplementation`.

`OcamlInterface` is the language used for `.mli` files, its associated
grammar is the OCaml interface grammar.
`OcamlImplementation` is the language used for `.ml` files, its
associated grammar is the OCaml implementation grammar.
`Ocaml` isn't associated with any file extension, and is associated with
both implementaiton and interface grammars, in that order.

All three languages are associated with the `ocaml.scm` query file. The
query file itself specifies that it uses the `Ocaml` language.

In practice, it means that if a `.ml` or `.mli` file is provided with
the `--input-file` argument, or if the `--language` argument is set to
`ocaml-implementation` or `ocaml-interface`, then the correct grammar
will be used.

However, if the input is provided via `stdin` and the `--language`
argument is `ocaml`, or the `--query` argument is `languages/ocaml.scm`,
then Topiary will try the implementation grammar, then the interface
grammar if the first fails. It will then format the resulting tree
with `ocaml.scm`.