Starting over to support comments
Because comments can trail other statements on the same line, it makes
sense to @append_hardline. Doing that however goes against the grain of
much of the rest of the queries, which _prepend_ new lines. As such, to
avoid comment queries being too difficult (or impossible?) to write, the
line feed discipline needs to be switched to append...

This is going to get ugly :(

[skip ci]