[Bash] Reverse new line strategy (#193)
Because comments can trail other statements on the same line, it makes sense to @append_hardline. Doing that however goes against the grain of much of the rest of the queries, which _prepend_ new lines. As such, to avoid comment queries being too difficult (or impossible?) to write, the line feed discipline needs to be switched to append...

* Semi-naively reverse the direction of new lines
* First approximation of formatting rules (in English) for comments
* Fix formatting for loops
* Fix formatting for case statements
* Fix formatting for if statements
* Improved command spacing and fixed formatting for function definitions
* Fix formatting for variables
* Fix spacing on single-line compounds, subshells and case branches
* Correct blank lines after heredoc end marker

NOTE A trailing new line exists in the output, as many nodes have @append_hardline regardless of context. These can be fixed with something like ((node) @append_hardline . (_)), but it's not a priority.