toml: allow inline comments in multiline arrays
previously, if you had comments on individual items in a multiline
array, e.g.

```toml
children = [
  "michael",
  "gob", # i don't care too much for this one
  "lindsay",
]
```

topiary would confusingly move the comment _below_ the element it was
placed next to-

```toml
children = [
  "michael",
  "gob",
  # i don't care too much for this one
  "lindsay",
]
```
which can destroy the intent of the comment. this rule update now
preserves inline comments and won't insert a linebreak after a separator
if followed by one.