toml: enforce final (oxford) comma in multiline arrays, remove from singleline arrays
this is more a preference, but it comes from a learned (python)
annoyance- if you want to keep a set of things in sorted order, then
every time the last item changes you have to go in and pick around the
commas. in python it's a lot worse due to implicit string coercion,
this is thankfully just illegal in toml, but still feels a good move for
consistency and ease of sorting.