Support Nickel patterns
The last (yet to be released) version of the Nickel grammar was extended
with enum variants and their corresponding patterns, and patterns can
now appear on the left-hand side of a match case (instead of just enum
tags). This commit updates Topiary to use the latest Nickel grammar and
update the Nickel queries to handle enum variants and patterns.

This commit also adds a scope for the body of a match case (the
right-hand side of "=>") for consistency reason, to align with how bound
expressions (right-hand side of "=") are flowed.