Good state
The modification of another query finally solves the indentation
problem. This query was dubious, because annotations are a series of
`(annot_atom)`, which most of the time fit on one line. We still want to
indent them, so we should have indented the whole series instead of each
node individually - which this commit is doing.

Also, the `multi_line_scop_only` is useless because if the annotation is
single line then the indentation just doesn't have any effect. Still, I
found that removing it is necessary to get the right behavior, this I
don't understand.