Update patch version (4.0.2 -> 4.0.3)
A previous merge commit added support for let-blocks to Nickel
formatting, a new syntactic construct. To make use of this update from
Nickel, we need to publish a new version of `topiary-queries`. As now
customary, we bump the patch version of the workspace, and only bother
publishing a new version of topiary-queries.