Nickel: formatting of single-line annotated let
Previously, we were formatting a let-binding followed by a single-line
annotation and some content by always putting the `=` sign and the
following content on a new line, indenting the content one level to the
right.

This was mostly due to a limitation of Topiary around scopes. Today,
this limitation has been lifted thanks to the introduction of measuring
scopes. We make use of the latter to finally format such let-bindings
(and bindings in general) as we originally wanted (see
https://github.com/tweag/topiary/issues/743).