Bump tree-sitter to 0.24 (not a release) (#794)
* Bump tree-sitter to 0.24

Biggest difficulty is that `tree_sitter::QueryMatches` is now a
`StreamingIterator`, returning references rather than values. This
makes it very difficult to store in a new type.

Instead make `topiary_tree_sitter_facade::QueryMatch` a trait rather
than a new type. This seems to be the easiest way to preserve as much
of the existing structure as possible.

Not tested for the wasm32 build, so possibly some changes required
there to match.

* Fix dynamically loading tree-sitter language functions

* Fix formatting

* Fix doctest comment

* chore: Mention `tree-sitter` bump in changelog